Refactor PodResourcesService
