+++ packages/app/services/pod-handling/pod-resources.js
@@ -1,11 +1,9 @@
-const { Readable } = require('stream');
-const { FormData } = require('formdata-node');
-const { FormDataEncoder } = require('form-data-encoder');
+const FetchPodOrProxyMixin = require('../../mixins/fetch-pod-or-proxy');
 const SparqlGenerator = require('sparqljs').Generator;
-const { stream2buffer } = require('../../utils');
 
 module.exports = {
   name: 'pod-resources',
+  mixins: [FetchPodOrProxyMixin],
   started() {
     this.sparqlGenerator = new SparqlGenerator({
       /* prefixes, baseIRI, factory, sparqlStar */
@@ -126,56 +124,6 @@ module.exports = {
         method: 'DELETE',
         actorUri
       });
-    },
-    async fetch(ctx) {
-      let { url, method, headers = {}, body, actorUri } = ctx.params;
-
-      const app = await ctx.call('app.get');
-
-      // True if the resource is not on the pod of the actor
-      const isLocal = url === actorUri || url.startsWith(actorUri + '/');
-
-      if (isLocal) {
-        return await ctx.call('signature.proxy.query', {
-          url,
-          method,
-          headers,
-          body,
-          actorUri: app.id || app['@id']
-        });
-      } else {
-        // Remote resources. We will go through the Pod proxy.
-        const actor = await ctx.call('activitypub.actor.get', { actorUri });
-        const proxyUrl = actor.endpoints?.proxyUrl;
-        if (!proxyUrl) throw new Error(`No proxy endpoint found for actor ${actorUri}`);
-
-        // Convert Headers object if necessary (otherwise we can't stringify it below)
-        // Note: if we use NodeJS built-in Headers instead of node-fetch Headers, the constructor name is _Headers
-        if (
-          headers &&
-          typeof headers === 'object' &&
-          (headers.constructor.name === 'Headers' || headers.constructor.name === '_Headers')
-        ) {
-          headers = Object.fromEntries(headers);
-        }
-
-        const formData = new FormData();
-
-        formData.append('id', url);
-        formData.append('method', method || 'GET');
-        formData.append('headers', JSON.stringify(headers));
-        if (body) formData.append('body', body);
-
-        const encoder = new FormDataEncoder(formData);
-
-        return await ctx.call('signature.proxy.query', {
-          url: proxyUrl,
-          method: 'POST',
-          headers: encoder.headers,
-          body: await stream2buffer(Readable.from(encoder)),
-          actorUri: app.id || app['@id']
-        });
-      }
     }
   }
 };
